Data cleaning
SQL data cleaning 10-22-2021

I noted that the columns contained data that appeared to be grossly incorrect, even though it was pulled from oracle's elixer which looked pretty clean.

I loaded the data into excel and noted by transposing column heads that about 8 columns had been added in 2020 that were not in earlier tables. Those tables had not been updated with those column values and so it was moving the data over.
I moved all of the new columns to the end and adjusted my SQL table creation to add in those columns at the end. This worked great except that when I saved it as an excel file briefly it somehow changed how the data was encoded.

I couldn't load the data into SQL because it wasn't UTF-8 encoded. Notepad++ had the file encoded as UTF-8 but showed there were some erroneous characters. Ended up having to write a script that decoded the file via python.