Dash Generation
I have some leads on how to deploy a dash server so I'm settling on that as my graph generator of choice for the moment.
https://datasciencecampus.github.io/deploy-dash-with-gcp/

I also checked what happened when I added the OPL stats and it moved none of the statistics appreciably so I'm leaving them out of the initial analysis for now.