Dash App post-mortem
I think I'm going to abandon using dash-app for now. If the goal here is to learn productivity skills that can be translated to real work then I probably want to be using Tableau, even if it's "harder"

I have the dash test folder setup from this: https://github.com/datasciencecampus/deploy-dash-with-gcp/tree/master/simple-dash-app-engine-app

But for now I'm not going to bother setting up a GCP for it.